Completed simple conversion. Added (hopefully) helpful README

+++ lib/guard/markdown.rb
@@ -8,32 +8,43 @@ module Guard
     # Your code goes here...
     def initialize(watchers=[], options={})
       super              
-      # init stuff here, thx!
-    end      
+      @options = {
+        :convert_on_start => true,
+        :dry_run          => false
+      }.update(options)
+    end         
+    
+    def start
+      UI.info("Guard::Markdown has started watching your files")
+      run_all if @options[:convert_on_start]
+    end
     
     def run_all
       files = Dir.glob("**/*.*")
-      targets = Watcher.match_files(self, files) 
-      #puts "Running changes on these paths: #{targets}" 
+      targets = Watcher.match_files(self, files)  
       run_on_change targets
     end
     
     # Called on file(s) modifications
+    # TODO  - this method does far too much. Must refactor to allow
+    #       - for better testing
     def run_on_change(paths)
       paths.each do |path|
         input, output = path.split("|") 
-        puts "#{input} >> #{output}"
-        source = File.open(input,"rb").read 
+        UI.info "#{input} >> #{output}" 
+        unless @options[:dry_run]
+          source = File.open(input,"rb").read 
                                             
-        # make sure directory path exists
-        reg = /(.+\/).+\.\w+/i
-        target_path = output.gsub(reg,"\\1")
-        FileUtils.mkpath target_path unless target_path.empty?
+          # make sure directory path exists
+          reg = /(.+\/).+\.\w+/i
+          target_path = output.gsub(reg,"\\1")
+          FileUtils.mkpath target_path unless target_path.empty?
         
-        doc = Kramdown::Document.new(source, :input => "markdown").to_html
+          doc = Kramdown::Document.new(source, :input => "markdown").to_html
         
-        File.open(output, "w") do |f|
-          f.puts(doc)
+          File.open(output, "w") do |f|
+            f.puts(doc)
+          end
         end
       end
       true
